requirements.txt
amqp==2.5.0
billiard==3.6.0.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
cloudinary==1.17.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-celery-beat==1.5.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-mathfilters==0.4.0
django-timezone-field==3.0
dnspython==1.16.0
eventlet==0.25.0
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
greenlet==1.1.0
gunicorn==20.0.4
httplib2==0.13.1
idna==2.8
kombu==4.6.3
mock==3.0.5
monotonic==1.5
oauthlib==3.1.0
pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
PyJWT==1.7.1
mysqlclient==1.4.6
python-crontab==2.3.8
python-dateutil==2.8.0
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.2
raven==6.10.0
redis==3.3.6
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
whitenoise==4.1.3
